Move ParameterValue class members to a dedicated .cpp files to avoid false-positive from CSA
